+++ R/bounded_niche.R
@@ -6,12 +6,15 @@ bounded_niche = function(g_min, g_max){
   #' @param g_min lowest value of the gradient the taxon can tolerate
   #' @param g_max highest value of the gradient the taxon can tolerate
   #'
-  #' @seealso [snd_niche()] for an alternative niche model, [apply_niche()] for the function that uses the function returned
+  #' @seealso
+    #' * [snd_niche()] for an alternative niche model
+    #' * [apply_niche()] for the function that uses the function returned
+    #' * `vignette("advanced_functionality")` for details how to create user-defined niche models
   #'
   #' @returns a function describing the niche for usage with `apply_niche`. The function returns 1 if the taxon is within its niche (the gradient is between `g_min` and `g_max`), and 0 otherwise
   #'
   #' @description
-    #' Defines a simple niche model where the niche defined is given by a lower limit (`g_min`) and an upper limie (`g_max`) of a gradient the taxon can tolerate
+    #' Defines a simple niche model where the niche defined is given by a lower limit (`g_min`) and an upper limit (`g_max`) of a gradient the taxon can tolerate
     #'
   #' @examples
     #'