+++ inst/doc/event_data.R
@@ -8,14 +8,14 @@ knitr::opts_chunk$set(
 library(StratPal)
 library(admtools)
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="histogram of fossil occurrences appearing at a constant rate."----
 # simulate fossil occurrences over one Myr with an average of 15 occurrences per Myr
 p3(rate = 15, from = 0, to = 1) |>
   hist(main = "Fossil abundance",
        xlab = "Time [Myr]",
        ylab = "# Specimens")
 
-## -----------------------------------------------------------------------------
+## ----fig.alt = "histogram of fossil occurrences appearing at a sinusoidal rate."----
 # return 100 occurrences by setting n parameter
 # note that negative rates (where sin < 0) are ignored
 p3_var_rate(x = sin, from = 0, to = 9, n = 100) |>
@@ -23,14 +23,14 @@ p3_var_rate(x = sin, from = 0, to = 9, n = 100) |>
        ylab= "# Specimens",
        main = "Fossil abundance")
 
-## -----------------------------------------------------------------------------
+## ----fig.alt = "histogram of fossil occurrences droping to 0."----------------
 # decline in last occurrences from 50 to 0 over 1 Myr
 p3_var_rate(x = c(0,1), y = c(50, 0), from = 0, to = 1, f_max = 50) |>
   hist(xlab = "Time [Myr]",
        main = "Last occurrences",
        ylab = "# Last occurrences")
 
-## ----figures-side, fig.show="hold", out.width="50%"---------------------------
+## ----figures-side, fig.show="hold", out.width="50%", fig.alt="2 plots of age-depth models, one 2 km from shore and one 12 km from shore."----
 adm_2km = tp_to_adm(t = scenarioA$t_myr,   # 2 km from shore
                 h = scenarioA$h_m[,"2km"],
                 T_unit = "Myr",
@@ -54,7 +54,7 @@ T_axis_lab()
 L_axis_lab() 
 title("Age-depth model 12 km from shore")
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="stratigraphic expression of fossil occurrences 2 km from shore."----
 p3(rate = 200, from = min_time(adm_2km), to = max_time(adm_2km)) |> # constant rate in time domain
   time_to_strat(adm_2km, destructive = TRUE) |>                     # transform into depth domain
   hist(xlab = "Stratigraphic height [m]",                           # plot
@@ -62,7 +62,7 @@ p3(rate = 200, from = min_time(adm_2km), to = max_time(adm_2km)) |> # constant r
        ylab = "# Fossils",
        breaks = seq(from = min_height(adm_2km), to = max_height(adm_2km), length.out = 20))
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="stratigraphic occurrence of fossil occurrences 12 km from shore."----
 p3(rate = 200, from = min_time(adm_12km), to = max_time(adm_12km)) |>  # same rate as 2 km from shore
   time_to_strat(adm_12km, destructive = TRUE) |>                       # use different adm for transformation
   hist(xlab = "Stratigraphic height [m]",                              # plot histogram
@@ -70,7 +70,7 @@ p3(rate = 200, from = min_time(adm_12km), to = max_time(adm_12km)) |>  # same ra
        ylab = "# Fossils",
        breaks = seq(from = min_height(adm_12km), to = max_height(adm_12km), length.out = 20))
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="stratigraphic expression of last occurrences 2 km from shore."----
 p3(rate = 200, from = min_time(adm_2km), to = max_time(adm_2km)) |> # constant rate of last occ
   time_to_strat(adm_2km, destructive = FALSE) |>                    # non-destructive transformation!
   hist(xlab = "Stratigraphic height [m]",                           # plot histogram
@@ -78,7 +78,7 @@ p3(rate = 200, from = min_time(adm_2km), to = max_time(adm_2km)) |> # constant r
        ylab = "# Last occurrences",
        breaks = seq(from = min_height(adm_2km), to = max_height(adm_2km), length.out = 20))
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="stratigraphic expression of last occurrences 12 km from shore."----
 p3(rate = 200, from = min_time(adm_12km), to = max_time(adm_12km)) |>
   time_to_strat(adm_12km, destructive = TRUE) |>
   hist(xlab = "Stratigraphic height [m]",
@@ -86,7 +86,7 @@ p3(rate = 200, from = min_time(adm_12km), to = max_time(adm_12km)) |>
        ylab = "# Last occurrences",
        breaks = seq(from = min_height(adm_12km), to = max_height(adm_12km), length.out = 20))
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="fossil occurrences in the time domain."-------------------------
 t_ext = 1.5 # time of "true" extinction
 r = 30      # rate of fossil occurrences
 # simulate rate fossil occurrences of taxon 
@@ -115,7 +115,7 @@ t_last_occ = highest_occ |> # time when last preserved fossil lived
 # time offset between true extinction and time when last fossil lived.
 time_range_offset_myr = t_ext - t_last_occ
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="water depth 2 km from shore in scenario A."---------------------
 t = scenarioA$t_myr           # time steps of the model
 wd = scenarioA$wd_m[,"2km"]   # water depth 2 km offshore at model time steps
 gc = approxfun(t, wd)         # define function that defines how the gradient changes with time (gc = *G*radient *C*hange)
@@ -126,7 +126,7 @@ plot(t, gc(t),
      ylab = "Water depth [m]",
      main = "Water depth 2 km offshore")
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="collection probability of a taxon along water depth."-----------
 my_niche = snd_niche(opt = 10,       # preferred water depth 
                      tol = 5,        # tolerance to water depth fluctuations
                      cutoff_val = 0) # cut off negative values - the taxon does not survive on land
@@ -137,7 +137,7 @@ plot(x, my_niche(x),
      ylab = "Collection probability",
      main = "Collection probability of taxon")
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="fossil abundance in the time domain, with incorporation of niche modeling."----
 p3(rate = 300, from = min_time(adm_2km), to = max_time(adm_2km)) |> # model occurrences of taxon based on a constant rate
   apply_niche(niche_def = my_niche, gc = gc) |>                     # apply the niche model
   hist(xlab = "Time [Myr]",
@@ -145,7 +145,7 @@ p3(rate = 300, from = min_time(adm_2km), to = max_time(adm_2km)) |> # model occu
        ylab = "# Fossils",
        breaks = seq(from = min_time(adm_2km), to = max_time(adm_2km), length.out = 40))
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="fossil abundance in the stratigraphic domain, with incorporation of niche modeling."----
 p3(rate = 300, from = min_time(adm_2km), to = max_time(adm_2km)) |> # model occurrences based on constant rate
   apply_niche(niche_def = my_niche, gc = gc) |>                     # apply niche model
   time_to_strat(adm_2km, destructive = TRUE) |>                     # transform into strat. domain, destroy fossils that coincide with hiatuses 
@@ -155,7 +155,7 @@ p3(rate = 300, from = min_time(adm_2km), to = max_time(adm_2km)) |> # model occu
        breaks = seq(from = min_height(adm_2km), to = max_height(adm_2km), length.out = 40))
 
 
-## -----------------------------------------------------------------------------
+## ----fig.alt="water depth 2 km from shore expressed in the stratigraphic domain."----
 list("t" = t, "y" = wd) |>    # create list with time - water depth information
   time_to_strat(adm_2km) |>   # transform into the strat. domain
   plot(orientation = "lr",    # plot water depth information in the stratigraphic domain
@@ -166,6 +166,7 @@ list("t" = t, "y" = wd) |>    # create list with time - water depth information
 
 ## ----eval=FALSE---------------------------------------------------------------
 #  vignette("phenotypic_evolution")
+#  vignette("paleoTS_functionality")
 
 ## ----eval=FALSE---------------------------------------------------------------
 #  vignette("advanced_functionality")